requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pillow>=10.0.0

# Web framework
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
